% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_letter_words.R
\name{n_letter_words}
\alias{n_letter_words}
\title{Create n-letter words}
\usage{
n_letter_words(
  n,
  num_letters = 26,
  case = c("upper", "lower"),
  as_vector = FALSE
)
}
\arguments{
\item{n}{the length of the words}

\item{num_letters}{how many letters to use in the word (an integer between 1 and 26)}

\item{case}{choose between upper or lower case letters}

\item{as_vector}{logical}
}
\value{
If as_vector = FALSE, a one column tibble with colname "word" containing all words of length n made with a combination of num_letter letters. If as_vector = TRUE, a vector of all such words.
}
\description{
Create all words of length n from given number of letters
}
\examples{
n_letter_words(3)
n_letter_words(4, num_letters = 2, case = "lower", as_vector = TRUE)
}
\author{
Ella Kaye
}
