% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digraph.R
\name{pdf_digraph}
\alias{pdf_digraph}
\title{Include digraphs in pdf output of Rmarkdown}
\usage{
pdf_digraph(filename, code)
}
\arguments{
\item{filename}{String of filename to save to, with exention .pdf}

\item{code}{Either a string defining a digraph in the DOT language, or a string of a .gv filename (this file must be stored in the same working directory).}
}
\value{
Plots a digraph in the pdf output of an Rmarkdown document, when function is embedded in a code chunk.
}
\description{
Include digraphs in pdf output of Rmarkdown
}
\examples{
\dontrun{
pdf_digraph_gv("graph.pdf", "graph.gv")
}
}
