% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digraph.R
\name{pdf_digraph_gv}
\alias{pdf_digraph_gv}
\title{Include digraphs in pdf output of Rmarkdown}
\usage{
pdf_digraph_gv(filename, gv)
}
\arguments{
\item{filename}{String of filename to save to, with exention .pdf}

\item{gv}{String of a .gv filename. This file must be stored in the same working directory}
}
\value{
Plots a digraph in the pdf output of an Rmarkdown document, when function is embedded in a code chunk.
}
\description{
Include digraphs in pdf output of Rmarkdown
}
\examples{
\dontrun{
pdf_digraph_gv("graph.pdf", "graph.gv")
}
}
