% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_vec.R
\name{same_elements}
\alias{same_elements}
\title{Test if all elements of a vector are equal}
\usage{
same_elements(x)
}
\arguments{
\item{x}{a vector}
}
\value{
logical
}
\description{
Test if all elements of a vector are equal
}
\seealso{
\link{zero_range}
}
