% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_vec.R
\name{zero_range}
\alias{zero_range}
\title{Test if all elements of a numeric vector are equal}
\usage{
zero_range(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric vector}

\item{tol}{the tolerance}
}
\value{
logical
}
\description{
Test if all elements of a numeric vector are equal
}
\seealso{
\link{same_elements}
}
