% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpres.R
\name{ggpres}
\alias{ggpres}
\title{Adjust test size on a ggplot to make it visible in presentations}
\usage{
ggpres(g, title_size = 24, text_size = 18)
}
\arguments{
\item{g}{A \code{ggplot} graph}

\item{title_size}{The font size for the title, legend title and axes titles}

\item{text_size}{The font size for for the axis lables, legend text and facet text}
}
\value{
A \code{ggplot} graph
}
\description{
Change the size of text on a \code{ggplot} so that a plot designed to look good in print is readable when included in a slide in a presentation (or vice verse)
}
\examples{

library(ggplot2)
g <- ggplot(mpg, aes(displ, hwy)) +
 geom_point() +
 facet_wrap(~class, scales = "free") +
 ggtitle("A faceted plot")
g2 <- ggpres(g)
g2 + geom_point(size = 2)
}
\author{
Ella Kaye
}
